refactor: move user to context and keep it mocked

+++ api/worklog.api.ts
@@ -5,28 +5,30 @@ import {
   createErrorResponse,
   createSuccessResponse,
 } from "#lib/responseFormatter";
-import { getRequiredEnvVar } from "#lib/env";
 
 const worklogService = new WorklogService();
 const router = createRouter();
-
-const userId = getRequiredEnvVar("MOCK_USER_ID");
 
 // Start work
 router.post("/api/worklogs/start", async (context) => {
   try {
     const body = (await context.request.json()) as unknown;
-    const input = z
+
+    const { description, location, startTime } = z
       .object({
         description: z.string().optional(),
         location: z.nativeEnum(WorkLocation),
         startTime: z.coerce.date(),
       })
       .parse(body);
+
     const worklog = await worklogService.startWork({
-      ...input,
-      userId,
+      userId: context.user.id,
+      startTime,
+      location,
+      description,
     });
+
     return createSuccessResponse(worklog);
   } catch (error) {
     return createErrorResponse(error);
@@ -38,18 +40,22 @@ router.post(
   "/api/worklogs/:id/complete",
   async (context: RouterContext<{ id: string }>) => {
     try {
-      const { id } = context.params;
+      const userId = context.user.id;
+      const worklogId = context.params.id;
       const body = (await context.request.json()) as unknown;
-      const input = z
+
+      const { endTime } = z
         .object({
           endTime: z.coerce.date(),
         })
         .parse(body);
+
       const worklog = await worklogService.completeWork({
-        worklogId: id,
+        worklogId,
         userId,
-        ...input,
+        endTime,
       });
+
       return createSuccessResponse(worklog);
     } catch (error) {
       return createErrorResponse(error);
@@ -62,20 +68,26 @@ router.patch(
   "/api/worklogs/:id",
   async (context: RouterContext<{ id: string }>) => {
     try {
-      const { id } = context.params;
+      const userId = context.user.id;
+      const worklogId = context.params.id;
       const body = (await context.request.json()) as unknown;
-      const input = z
+
+      const { startTime, description, location } = z
         .object({
           startTime: z.coerce.date().optional(),
           description: z.string().optional(),
           location: z.nativeEnum(WorkLocation).optional(),
         })
         .parse(body);
+
       const worklog = await worklogService.modifyWork({
-        ...input,
-        worklogId: id,
+        worklogId,
         userId,
+        startTime,
+        description,
+        location,
       });
+
       return createSuccessResponse(worklog);
     } catch (error) {
       return createErrorResponse(error);
@@ -86,8 +98,10 @@ router.patch(
 // Create completed worklog
 router.post("/api/worklogs", async (context) => {
   try {
+    const userId = context.user.id;
     const body = (await context.request.json()) as unknown;
-    const input = z
+
+    const { startTime, endTime, description, location } = z
       .object({
         startTime: z.coerce.date(),
         endTime: z.coerce.date(),
@@ -95,10 +109,15 @@ router.post("/api/worklogs", async (context) => {
         location: z.nativeEnum(WorkLocation),
       })
       .parse(body);
+
     const worklog = await worklogService.createCompletedWork({
-      ...input,
       userId,
+      startTime,
+      endTime,
+      description,
+      location,
     });
+
     return createSuccessResponse(worklog);
   } catch (error) {
     return createErrorResponse(error);
@@ -108,16 +127,17 @@ router.post("/api/worklogs", async (context) => {
 // Delete worklog
 router.delete("/api/worklogs", async (context) => {
   try {
+    const userId = context.user.id;
     const body = (await context.request.json()) as unknown;
-    const input = z
+
+    const { worklogId } = z
       .object({
         worklogId: z.string(),
       })
       .parse(body);
-    await worklogService.deleteWork({
-      ...input,
-      userId,
-    });
+
+    await worklogService.deleteWork({ worklogId, userId });
+
     return createSuccessResponse({ success: true });
   } catch (error) {
     return createErrorResponse(error);
@@ -129,11 +149,11 @@ router.get(
   "/api/worklog/:id",
   async (context: RouterContext<{ id: string }>) => {
     try {
-      const { id } = context.params;
-      const worklog = await worklogService.getWorklog({
-        worklogId: id,
-        userId,
-      });
+      const userId = context.user.id;
+      const worklogId = context.params.id;
+
+      const worklog = await worklogService.getWorklog({ worklogId, userId });
+
       return createSuccessResponse(worklog);
     } catch (error) {
       return createErrorResponse(error);
@@ -144,17 +164,24 @@ router.get(
 // Query worklogs
 router.get("/api/worklogs", async (context) => {
   try {
-    const input = z
+    const userId = context.user.id;
+    const params = context.url.searchParams;
+
+    const { startDate, endDate, status } = z
       .object({
         startDate: z.coerce.date().optional(),
         endDate: z.coerce.date().optional(),
         status: z.nativeEnum(WorklogStatus).optional(),
       })
-      .parse(context.url.searchParams);
+      .parse(params);
+
     const worklogs = await worklogService.getUserWorks({
-      ...input,
       userId,
+      startDate,
+      endDate,
+      status,
     });
+
     return createSuccessResponse(worklogs);
   } catch (error) {
     return createErrorResponse(error);
